sesdev: split box-related code into new sesdev/box.py
". . . its ongoing mission to make the sesdev codebase less
spaghetti-like . . ."
